L4/Project Building a Multi-Agent AI-Powered Mock Interviewer using CrewAI/requirements.txt
# AI Mock Interviewer — Python dependencies
# NO OpenAI. NO CrewAI LLM wrappers. NO LiteLLM.
# LLMs called directly: Ollama (requests) + Groq (groq SDK).

# ── UI ─────────────────────────────────────────────────
streamlit>=1.35.0

# ── LLM providers ──────────────────────────────────────
groq>=0.9.0            # Groq cloud API (secondary)
# Ollama uses the built-in `requests` library (no extra package needed)

# ── Utilities ──────────────────────────────────────────
python-dotenv>=1.0.0
requests>=2.31.0
